+++ home-manager/configs/root.nix
@@ -2,32 +2,20 @@
 ##                                                                            ##
 ## * Use Zsh                                                                  ##
 ## * Include custom Nix aliases (see `../zsh/nix.zsh`)                        ##
-## * Sources `/etc/zprofile` in case it has not been sourced before           ##
 ##                                                                            ##
 ################################################################################
 
 { config, pkgs, ... }:
 
 {
-  home.file = {
-    # Zsh aliases and environments
-    ".zsh/aliases.zsh".source = ../zsh/aliases.zsh;
-    ".zsh/nix.zsh".source = ../zsh/nix.zsh;
+  imports = [ ./.. ];
+
+  confkit.zsh = {
+    enable = true;
+    plugins = [ "aliases" "nix" ];
   };
 
   programs.home-manager = {
     enable = true;
-  };
-
-  programs.zsh = {
-    enable = true;
-
-    initExtra = ''
-      source /etc/zprofile
-      for script ($HOME/.zsh/*.zsh); do
-        source $script
-      done
-      unset script
-    '';
   };
 }