+++ nixos/default.nix
@@ -1,0 +1,12 @@
+{
+  imports = [
+    ../common/keyboard.nix
+    ./modules/nix.nix
+    ./modules/ranger.nix
+    ./modules/shell.nix
+    ./modules/tmux.nix
+    ./modules/utilities.nix
+    ./modules/vim.nix
+    ./modules/zsh.nix
+  ];
+}