+++ zsh/xen.zsh
@@ -1,0 +1,14 @@
+###############
+# Xen Aliases #
+###############
+
+alias x='sudo xl'
+alias xi='sudo xl info'
+alias xls='sudo xl list'
+alias xt='sudo xl top'
+alias xc='sudo xl create'
+alias xcc='sudo xl create -c'
+alias xco='sudo xl console'
+alias xr='sudo xl reboot'
+alias xs='sudo xl shutdown'
+alias xd='sudo xl destroy'